Palette code does not need to include actual tweaks!
