+++ docs/assets/scripts/tweak.js
@@ -1,6 +1,6 @@
 /**
  * Get import code for remixed themes and tweaked palettes.
  */
-export { palette as getPaletteCode, theme as getThemeCode } from './tweak/code.js';
+export { theme as getThemeCode } from './tweak/code.js';
 export { cdnUrl, hueRanges, hues, selectors, tints, urls } from './tweak/data.js';
 export { default as Permalink } from './tweak/permalink.js';
